Backport 3c276ce1fec17ed0238cc977320f2c82965fc45b
